refactor(frontend): extract admin dashboard cluster from main.ts (#312)
Conservative partial split (per the issue's lower-risk option): pull the
self-contained admin surface — admin dashboard (#241e: stats, cross-user
sessions, force stop/delete, edit caps), admin Groups CRUD, admin
observe-log, and the per-session Observers modal — out of the 5229-line
main.ts into admin.ts (~930 lines). Also extracts the three pure formatters
shared by the session list and the admin dashboard into a leaf format.ts.

Circular-import safety: admin.ts re-queries its DOM elements locally via
getElementById (elements are singletons by id), so it has NO module-top
dependency on main.ts — avoiding the TDZ trap. Its only main.ts imports
(showToast, activeSessionId) are used inside functions, resolving at call
time after both modules load. The 5 shared modal consts the core Escape
guard / applyAdminVisibility reference stay in main.ts. main.ts imports the
four close* functions its modal-registry code calls back into.

Behaviour-preserving: handler bodies + wirings moved verbatim. The admin
event wirings now register when admin.ts is imported (still after DOM parse,
since the entry script is a deferred module) — addEventListener order across
files is not observable.

main.ts: 5229 → 4324 lines. tsc + Biome clean, vite build green, frontend
tests pass (66). Remaining modal features (new-session, ports, invites,
paste, my-groups) follow the same pattern and can be extracted incrementally.
